+++ examples/ads_pixel_events_post_custom.js
@@ -11,6 +11,7 @@
 const bizSdk = require('facebook-nodejs-business-sdk');
 const Content = bizSdk.Content;
 const CustomData = bizSdk.CustomData;
+const DeliveryCategory = bizSdk.DeliveryCategory;
 const EventRequest = bizSdk.EventRequest;
 const UserData = bizSdk.UserData;
 const ServerEvent = bizSdk.ServerEvent;
@@ -22,14 +23,21 @@ const api = bizSdk.FacebookAdsApi.init(access_token);
 let current_timestamp = Math.floor(new Date() / 1000);
 
 const userData = (new UserData())
-                .setEmail('joe@eg.com')
+                .setEmails(['joe@eg.com'])
+                .setPhones(['12345678901', '14251234567'])
                 // It is recommended to send Client IP and User Agent for Conversions API Events.
                 .setClientIpAddress(request.connection.remoteAddress)
                 .setClientUserAgent(request.headers['user-agent'])
                 .setFbp('fb.1.1558571054389.1098115397')
                 .setFbc('fb.1.1554763741205.AbCdEfGhIjKlMnOpQrStUvWxYz1234567890');
 
+const content = (new Content())
+                .setId('product123')
+                .setQuantity(1)
+                .setDeliveryCategory(DeliveryCategory.HOME_DELIVERY);
+
 const customData = (new CustomData())
+                .setContents([content])
                 .setCurrency('usd')
                 .setValue(123.45);