+++ src/objects/ad-study.js
@@ -10,6 +10,7 @@ import {AbstractCrudObject} from './../abstract-crud-object';
 import AbstractObject from './../abstract-object';
 import Cursor from './../cursor';
 import AdStudyCell from './ad-study-cell';
+import PrivateLiftStudyInstance from './private-lift-study-instance';
 import AdStudyObjective from './ad-study-objective';
 
 /**
@@ -55,6 +56,25 @@ export default class AdStudy extends AbstractCrudObject {
       params,
       fetchFirstPage,
       '/cells'
+    );
+  }
+
+  getInstances (fields: Array<string>, params: Object = {}, fetchFirstPage: boolean = true): Cursor | Promise<*> {
+    return this.getEdge(
+      PrivateLiftStudyInstance,
+      fields,
+      params,
+      fetchFirstPage,
+      '/instances'
+    );
+  }
+
+  createInstance (fields: Array<string>, params: Object = {}): Promise<PrivateLiftStudyInstance> {
+    return this.createEdge(
+      '/instances',
+      fields,
+      params,
+      PrivateLiftStudyInstance
     );
   }